Upload some changes.
1. Upload image sources.
2. Upload weight function class.

+++ mainwindow.cpp
@@ -20,7 +20,7 @@ MainWindow::~MainWindow()
 
 void MainWindow::nextWidget()
 {
-    if(curWidget == widgetList.size()-1){
+    if((int)curWidget == widgetList.size()-1){
         QApplication::exit(0);
     }else{
         widgetList.at(curWidget)->hide();
